moved common fixtures to conftest.py
