+++ tests/pygen/test_templates.py
@@ -1,36 +1,8 @@
 """Tests for the various features from the code generation templates."""
 import importlib
-import os
-import shutil
-
-import pytest
-import sys
 
 from pyecore.ecore import EPackage
-from pyecore.resources import ResourceSet, URI
 from pygen.ecore import EcoreGenerator
-
-
-@pytest.fixture('module', autouse=True)
-def cwd_module_dir():
-    # change cwd to this module's directory:
-    cwd = os.getcwd()
-    os.chdir(os.path.dirname(__file__))
-    yield
-
-    # reset after module goes out of scope:
-    os.chdir(cwd)
-
-
-@pytest.fixture
-def pygen_output_dir():
-    path = os.path.join('output', 'template_features')
-    shutil.rmtree(path, ignore_errors=True)
-    original_sys_path = sys.path
-    sys.path.append(path)
-    yield path
-    sys.path.remove(path)
-    #shutil.rmtree(path, ignore_errors=False)
 
 
 def generate_meta_model(model, output_dir):
